Move storage out of UI screen and add checks for parsing failures
